+++ packages/tizen_bundle/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## NEXT
 
 * Update code format.
-* Add 35 integration test cases.
+* Update integration test suite to 37 test cases.
 
 ## 0.1.2